requirements.txt
pandas
numpy

sentence-transformers
torch

faiss-cpu

rank-bm25

fastapi
uvicorn

pytest>=7.0
pytest-asyncio>=0.21.0
httpx>=0.24.0

scikit-learn
jupyter
matplotlib
seaborn
tf-keras